feat: Estruuração do processo de checkout

+++ src/Soat10.TechChallenge.Application/Entities/CartItem.cs
@@ -37,7 +37,7 @@ namespace Soat10.TechChallenge.Application.Entities
         public virtual Product Product { get; set; }
         public int Quantity { get; private set; }
         public string? Notes { get; private set; }
-        public double Price => (double)(Product?.Price);
+        public decimal Price => (decimal)(Product?.Price);
 
     }
 }
